pyproject.toml
[project]
name = "xtquant-grpc"
version = "0.1.0"
description = "gRPC wrapper for xtquant (MiniQMT Python SDK)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "pandas",
    "numpy",
]

[tool.hatch.build.targets.wheel]
packages = ["server", "pb"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["test"]
markers = [
    "slow: Long-running tests (financial data download, etc.), run with -m slow",
]
addopts = "-m 'not slow'"
